- move Data/PostObjectType to Registry/Utils/PostObject - move Data/TermObjectType to Registry/Utils/TermObject
