+++ src/Type/ObjectType/TermObject.php
@@ -2,6 +2,7 @@
 
 namespace WPGraphQL\Type\ObjectType;
 
+use Exception;
 use WP_Taxonomy;
 use WPGraphQL\Data\TermObjectType;
 
@@ -18,14 +19,15 @@ class TermObject {
 	 *
 	 * @param WP_Taxonomy $tax_object The taxonomy being registered
 	 *
+	 * @return void
+	 * @throws Exception
 	 * @deprecated @todo
 	 *
-	 * @return void
 	 */
 	public static function register_taxonomy_object_type( WP_Taxonomy $tax_object ) {
-		_deprecated_function( __FUNCTION__, '@todo', TermObjectType::class . '::register_term_object_types()' ); //phpcs:ignore WordPress.Security.EscapeOutput.OutputNotEscaped
+		_deprecated_function( __FUNCTION__, '@todo', \WPGraphQL\Registry\Utils\TermObject::class . '::register_term_object_types()' );
 
-		TermObjectType::register_term_object_types( $tax_object );
+		\WPGraphQL\Registry\Utils\TermObject::register_term_object_types( $tax_object );
 	}
 
 }
